working conversion

+++ src/utils/mp3_to_x.py
@@ -4,7 +4,7 @@ import scipy
 import numpy as np
 from tqdm import tqdm
 
-def convert_folder_to_extension(folder_path, new_folder = None, extension=  'wav'):
+def convert_folder_to_extension(folder_path, new_folder = None, extension=  'wav', delete = False):
     for root, dirs, files in os.walk(folder_path):
         if new_folder is None:
             folder = root
@@ -16,7 +16,7 @@ def convert_folder_to_extension(folder_path, new_folder = None, extension=  'wav
                 try:
                     new_name = name[:-3]+extension
                     if not os.path.exists(os.path.join(folder,new_name)):
-                        print(f"saving {os.path.join(root,name)} to {os.path.join(folder,name)}")
+                        print(f"saving {os.path.join(root,name)} to {os.path.join(folder,new_name)}")
                         # y,sr =  librosa.load(os.path.join(root,name))
                         # if not os.path.exists(folder):
                         #     os.makedirs(folder)
@@ -27,13 +27,14 @@ def convert_folder_to_extension(folder_path, new_folder = None, extension=  'wav
                             
                 except:
                     print(os.path.join(root,name))
-                    os.remove(os.path.join(root,name))
+                    if delete:
+                        os.remove(os.path.join(root,name))
                 
 
 
-def conver_folder_to_wav(folder_path):
-    convert_folder_to_extension(folder_path=folder_path,extension='wav')
+def convert_folder_to_wav(folder_path, new_folder = None, delete = False):
+    convert_folder_to_extension(folder_path=folder_path,extension='wav', new_folder=new_folder, delete=delete)
 
 
-def conver_folder_to_npy(folder_path):
-    convert_folder_to_extension(folder_path=folder_path,extension='npy')
+def convert_folder_to_npy(folder_path, new_folder = None, delete = False):
+    convert_folder_to_extension(folder_path=folder_path,extension='npy', new_folder=new_folder, delete=delete)
